Tag v1.4.0: Merge everything for a read-only DB. Set Version to 1.4.0
Merged revision(s) 172-175 from branches/readonly_db:
Add a username / password parameter for a ready-only and one for read-write connection. Default to constructing a read-only database.
........
Add a paremter to test if a Database is read-only
........
Logger creates a new Database-instance if the passed one is read-only
........
Create writeable handlers where its needed
........